requirements-dev.txt
docker>=3.7.0,<3.8.0
exam>=0.5.1
freezegun==1.1.0
honcho>=1.0.0,<1.1.0
openapi-core @ https://github.com/getsentry/openapi-core/archive/master.zip#egg=openapi-core
<<<<<<< Updated upstream
pytest==6.1.0
pytest-cov==2.11.1
pytest-django==3.10.0
=======
pytest==4.6.5
pytest-cov==2.5.1
pytest-django==3.5.1
>>>>>>> Stashed changes
pytest-sentry==0.1.6
pytest-rerunfailures==9.1.1
responses>=0.8.1,<0.9.0
sqlparse==0.2.4
